requirements.txt
aiofiles==24.1.0
aioredis==2.0.1
aiosmtplib==4.0.0
aiosqlite==0.21.0
alembic==1.15.1
annotated-types==0.7.0
anyio==4.8.0
async-timeout==5.0.1
asyncpg==0.30.0
babel==2.17.0
base32hex==1.0.2
bcrypt==4.3.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cli_helpers==2.4.0
click==8.1.8
configobj==5.0.9
cryptography==44.0.2
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.11
fastapi-admin==1.0.4
fastapi-cli==0.0.7
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
iso8601==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
packaging==24.2
pendulum==3.0.0
pgcli==4.3.0
pgspecial==2.1.3
pinecone==6.0.2
pinecone-plugin-assistant==1.4.0
pinecone-plugin-interface==0.0.7
prompt_toolkit==3.0.50
psycopg==3.2.5
psycopg2-binary==2.9.10
pycparser==2.22
pycryptodome==3.21.0
pycryptodomex==3.21.0
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pypika-tortoise==0.5.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
rich-toolkit==0.13.2
setproctitle==1.3.5
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sqladmin==0.20.1
SQLAlchemy==2.0.38
sqlparse==0.5.3
starlette==0.46.1
tabulate==0.9.0
time-machine==2.16.0
tortoise-orm==0.24.2
typer==0.15.2
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3.1
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
wcwidth==0.2.13
websockets==15.0.1
WTForms==3.1.2
